pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "open-claw"
version = "0.1.0"
description = "AI agent workforce management system"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "DecawDevonn", email = "contact@openclaw.dev"}]
keywords = ["ai", "agents", "workforce", "automation", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "gunicorn>=21.0",
    "python-dotenv>=1.0.1",
    "PyJWT>=2.12.0",
    "cryptography>=46.0.5",
    "requests>=2.32.3",
    "redis>=5.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
]
mongo = [
    "pymongo>=4.0",
]
monitoring = [
    "sentry-sdk[flask]>=2.6.0",
]
memory = [
    "chromadb>=1.0.0",
]
all = [
    "pymongo>=4.0",
    "sentry-sdk[flask]>=2.6.0",
    "chromadb>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/DecawDevonn/open-claw"
Repository = "https://github.com/DecawDevonn/open-claw"
Issues = "https://github.com/DecawDevonn/open-claw/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "*.egg-info"]
